Move vm snapshot and restore logic to new storage pkg
with a goal to eventually move all the storage logic to
storage pkg.
