+++ pkg/storage/snapshot/util.go
