+++ pkg/virt-controller/watch/BUILD.bazel
@@ -27,6 +27,7 @@ go_library(
         "//pkg/monitoring/profiler:go_default_library",
         "//pkg/monitoring/vmistats:go_default_library",
         "//pkg/service:go_default_library",
+        "//pkg/storage/snapshot:go_default_library",
         "//pkg/util:go_default_library",
         "//pkg/util/cluster:go_default_library",
         "//pkg/util/lookup:go_default_library",
@@ -44,7 +45,6 @@ go_library(
         "//pkg/virt-controller/watch/drain/disruptionbudget:go_default_library",
         "//pkg/virt-controller/watch/drain/evacuation:go_default_library",
         "//pkg/virt-controller/watch/export:go_default_library",
-        "//pkg/virt-controller/watch/snapshot:go_default_library",
         "//pkg/virt-controller/watch/topology:go_default_library",
         "//pkg/virt-controller/watch/util:go_default_library",
         "//pkg/virt-controller/watch/workload-updater:go_default_library",
@@ -112,6 +112,7 @@ go_test(
         "//pkg/instancetype:go_default_library",
         "//pkg/monitoring/migration:go_default_library",
         "//pkg/rest:go_default_library",
+        "//pkg/storage/snapshot:go_default_library",
         "//pkg/testutils:go_default_library",
         "//pkg/util/types:go_default_library",
         "//pkg/virt-config:go_default_library",
@@ -120,7 +121,6 @@ go_test(
         "//pkg/virt-controller/watch/drain/disruptionbudget:go_default_library",
         "//pkg/virt-controller/watch/drain/evacuation:go_default_library",
         "//pkg/virt-controller/watch/export:go_default_library",
-        "//pkg/virt-controller/watch/snapshot:go_default_library",
         "//pkg/virt-controller/watch/topology:go_default_library",
         "//pkg/virt-controller/watch/util:go_default_library",
         "//staging/src/kubevirt.io/api/clone/v1alpha1:go_default_library",
